requirements.txt
fastapi
uvicorn[standard]
aiofiles
pydantic-settings
python-magic
gunicorn
python-multipart